reafactor: move controller schemas
