+++ 2-crud-api/src/index.ts
@@ -1,11 +1,16 @@
 import dotenv from 'dotenv';
 import express, { Express, Request, Response } from 'express';
-import { addUserSchema, AddUserSchema, patchUserSchema, PatchUserSchema, autoSuggestSchema, AutoSuggestSchema, getUserSchema, GetUserSchema } from './models';
 import { ValidatedRequest } from 'express-joi-validation';
 import {
   StatusCodes,
 } from 'http-status-codes';
 
+import {
+  getUserSchema, GetUserSchema,
+  addUserSchema, AddUserSchema,
+  patchUserSchema, PatchUserSchema,
+  autoSuggestSchema, AutoSuggestSchema,
+} from './controllers/user';
 import UserService from './services/user';
 import { pgInit } from './loaders/pgInit';
 
@@ -13,8 +18,14 @@ dotenv.config();
 
 async function start() {
   await pgInit();
+
+  const app: Express = express();
+  const port = process.env.PORT;
+  app.use(express.json());
+
   const userService = new UserService();
-  const Controller = {
+
+  const Router = {
     getUser: {
       route: '/',
       service: userService.get
@@ -37,46 +48,41 @@ async function start() {
     },
   };
 
-  const app: Express = express();
-  const port = process.env.PORT;
-
-  app.use(express.json());
-
-  app.get(Controller.getUser.route, getUserSchema, async (req: ValidatedRequest<GetUserSchema>, res: Response) => {
-    const { data, error } = await Controller.getUser.service(req.query);
+  app.get(Router.getUser.route, getUserSchema, async (req: ValidatedRequest<GetUserSchema>, res: Response) => {
+    const { data, error } = await Router.getUser.service(req.query);
     if (error) {
       return res.status(StatusCodes.INTERNAL_SERVER_ERROR).json(error);
     }
     res.status(StatusCodes.OK).json(data);
   });
 
-  app.post(Controller.addUser.route, addUserSchema, async (req: ValidatedRequest<AddUserSchema>, res: Response
+  app.post(Router.addUser.route, addUserSchema, async (req: ValidatedRequest<AddUserSchema>, res: Response
   ) => {
-    const { data, error } = await Controller.addUser.service(req.body);
+    const { data, error } = await Router.addUser.service(req.body);
     if (error) {
       return res.status(StatusCodes.INTERNAL_SERVER_ERROR).json(error);
     }
     res.status(StatusCodes.OK).json(data);
   });
 
-  app.patch(Controller.patchUser.route, patchUserSchema, async (req: ValidatedRequest<PatchUserSchema>, res: Response) => {
-    const { data, error } = await Controller.patchUser.service(req.body);
+  app.patch(Router.patchUser.route, patchUserSchema, async (req: ValidatedRequest<PatchUserSchema>, res: Response) => {
+    const { data, error } = await Router.patchUser.service(req.body);
     if (error) {
       return res.status(StatusCodes.INTERNAL_SERVER_ERROR).json(error);
     }
     res.status(StatusCodes.OK).json(data);
   });
 
-  app.delete(Controller.deleteUser.route, async (req: Request, res: Response) => {
-    const { data, error } = await Controller.deleteUser.service(req.body);
+  app.delete(Router.deleteUser.route, async (req: Request, res: Response) => {
+    const { data, error } = await Router.deleteUser.service(req.body);
     if (error) {
       return res.status(StatusCodes.INTERNAL_SERVER_ERROR).json(error);
     }
     res.status(StatusCodes.OK).json(data);
   });
 
-  app.get(Controller.getUsersList.route, autoSuggestSchema, async (req: ValidatedRequest<AutoSuggestSchema>, res: Response) => {
-    const { data, error } = await Controller.getUsersList.service(req.query);
+  app.get(Router.getUsersList.route, autoSuggestSchema, async (req: ValidatedRequest<AutoSuggestSchema>, res: Response) => {
+    const { data, error } = await Router.getUsersList.service(req.query);
     if (error) {
       return res.status(StatusCodes.INTERNAL_SERVER_ERROR).json(error);
     }
